Break Warc's dependency on db and scrub code
